requirements/dev.txt
# Specifies only dev-specific requirements
# But imports the common ones too
-r common.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-debug-toolbar==3.2
Pillow==8.0.1
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
